Build improvements and fixes
- Add ccache support for faster compilation
- Fix GCC multichar warning in reTuner target
- Fix enum bitwise-or warnings in RubberBandShifter (explicit casts)
- Add install targets for plugins and standalone app
- Add CPack configuration for creating installers
- Update copyright notice
- Remove artifact upload from CI workflow

+++ src/rubberbandshifter.hpp
@@ -171,10 +171,10 @@ private:
     void createOrReconfigureStretcher()
     {
         using RBS = RubberBand::RubberBandStretcher;
-        const size_t options = RBS::DefaultOptions
-                               | RBS::OptionProcessRealTime
-                               | RBS::OptionPitchHighConsistency
-                               | RBS::OptionThreadingNever;
+        const size_t options = (size_t) RBS::DefaultOptions
+                               | (size_t) RBS::OptionProcessRealTime
+                               | (size_t) RBS::OptionPitchHighConsistency
+                               | (size_t) RBS::OptionThreadingNever;
 
         auto sampleRate = static_cast<size_t> (juce::roundToInt (_sampleRate));
         auto channelCount = static_cast<size_t> (_numChannels);
